+++ command/token_capabilities_test.go
@@ -171,15 +171,45 @@ func TestTokenCapabilitiesCommand_Run(t *testing.T) {
 		client, closer := testVaultServer(t)
 		defer closer()
 
-		_, cmd := testTokenCapabilitiesCommand(t)
+		policy := `path "secret/foo" { capabilities = ["read"] } path "secret/bar" { capabilities = ["list"] }`
+		if err := client.Sys().PutPolicy("policy", policy); err != nil {
+			t.Error(err)
+		}
+
+		secret, err := client.Auth().Token().Create(&api.TokenCreateRequest{
+			Policies: []string{"policy"},
+			TTL:      "30m",
+		})
+		if err != nil {
+			t.Fatal(err)
+		}
+		if secret == nil || secret.Auth == nil || secret.Auth.ClientToken == "" {
+			t.Fatalf("missing auth data: %#v", secret)
+		}
+		token := secret.Auth.ClientToken
+
+		client.SetToken(token)
+
+		ui, cmd := testTokenCapabilitiesCommand(t)
 		cmd.client = client
 
 		code := cmd.Run([]string{
 			"secret/foo,secret/bar",
 		})
-		if exp := 1; code != exp {
+		if exp := 0; code != exp {
 			t.Errorf("expected %d to be %d", code, exp)
 		}
+
+		expectedFoo := "secret/foo    read"
+		expectedBar := "secret/bar    list"
+		combined := ui.OutputWriter.String() + ui.ErrorWriter.String()
+		if !strings.Contains(combined, expectedFoo) {
+			t.Errorf("expected %q to contain %q", combined, expectedFoo)
+		}
+		if !strings.Contains(combined, expectedBar) {
+			t.Errorf("expected %q to contain %q", combined, expectedBar)
+		}
+
 	})
 
 	t.Run("no_tabs", func(t *testing.T) {