+++ CHANGELOG.md
@@ -23,6 +23,8 @@ IMPROVEMENTS:
  * core/seal (enterprise): Lazily rewrap data when seal keys are rotated
  * expiration: Allow revoke-prefix and revoke-force to work on single leases as
    well as prefixes [GH-4450]
+ * cli: Display all capabilities when running `vault token capabilities` with
+   multiple paths [GH-4554]
 
 BUG FIXES: